fix: preserve legacy sign list time scalars

+++ internal/core/domain/economy.go
@@ -113,7 +113,7 @@ type SignInResult struct {
 
 type SignInListEntry struct {
 	UserID       string
-	SignedAtUnix int64
+	SignedAtUnix float64
 	ShowSignedAt bool
 }
